feat: add image downloading functionality
This commit adds a new method `download_image` to the `Scrapper` class. This method allows downloading images from the given URL and saving them with the specified filename. It iterates over the list of dictionaries containing the image URLs and uses the `requests` library to download the images. The downloaded images are saved as binary files using the `wb` mode.

The purpose of this change is to enhance the functionality of the `Scrapper` class by enabling the downloading of images along with the scraping of data.